Added some basic structure for the future CharacterSheet stuff.
The Controllers(under XF\Pub\Controller) basically do backend shenanigans for stuff.
CharacterSheet itself is a new Controller for the future CharacterSheet, currently not doing much.
The Member.php is an extension to the Member Controller of Xenforo, which basically has all controlling logic for the profile.
The character_sheet_view_template is the template to do the looks at the character sheets. I made it manually inside the output folder.

Naming conventions are basically to mirror things how they are for the original XenForo.